pyproject.toml
[tool.poetry]
name = "ecommerce-api"
version = "0.1.0"
description = ""
authors = ["eedu7 <mueedahmad067@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.114.2"}
sqlalchemy = "^2.0.34"
alembic = "^1.13.2"
pydantic-settings = "^2.5.2"
python-dotenv = "^1.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
aiomysql = "^0.2.0"
jinja2 = "^3.1.4"
bcrypt = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
